pyproject.toml
[tool.poetry]
name = "django-rest-sliknik"
version = "0.1.0"
description = ""
authors = ["Nikal Morgan <nikalmorgan11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"

[tool.poetry.dev-dependencies]
djangorestframework = "^3.11.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
